[daniel.tomi] 방탈출-authorization step1,2 리뷰 요청 드립니다! (#16)
* feat : member 테이블 username을 unique로 바꾸고 예외처리를 함

* feat : 로그인시 토큰 발급 기능 추가

* feat : 토큰을 이용하여 자기정보 조회

* feat : 예약 테스트 코드 수정

* feat : config 작성

* feat : 예약하기, 예약취소 개선

* fix : @AuthenticationPrincipal 적용 시 Member Entity가 아닌 LoginMember 형식의 객체를 반환하는 것으로 고침.

그 외 자잘한 docs작업을 하였음.

* docs : javadocs 추가

* docs : README.md에 요구사항 resolved 표시

* fix : @AuthorizationPrincipal에서 username이 아닌 memberId값을 반환하도록 함, 관련 클래스 고침

+++ src/main/java/nextstep/reservation/ReservationService.java
@@ -1,11 +1,13 @@
 package nextstep.reservation;
 
+import nextstep.member.MemberDao;
 import nextstep.schedule.Schedule;
 import nextstep.schedule.ScheduleDao;
 import nextstep.support.DuplicateEntityException;
 import nextstep.theme.Theme;
 import nextstep.theme.ThemeDao;
 import org.springframework.stereotype.Service;
+import nextstep.support.AuthorizationException;
 
 import java.util.List;
 
@@ -14,15 +16,18 @@ public class ReservationService {
     public final ReservationDao reservationDao;
     public final ThemeDao themeDao;
     public final ScheduleDao scheduleDao;
+    private final MemberDao memberDao;
 
-    public ReservationService(ReservationDao reservationDao, ThemeDao themeDao, ScheduleDao scheduleDao) {
+    public ReservationService(ReservationDao reservationDao, ThemeDao themeDao, ScheduleDao scheduleDao, MemberDao memberDao) {
         this.reservationDao = reservationDao;
         this.themeDao = themeDao;
         this.scheduleDao = scheduleDao;
+        this.memberDao = memberDao;
     }
 
-    public Long create(ReservationRequest reservationRequest) {
+    public Long create(ReservationRequest reservationRequest, Long memberId) {
         Schedule schedule = scheduleDao.findById(reservationRequest.getScheduleId());
+        String memberName = memberDao.findById(memberId).getUsername();
         if (schedule == null) {
             throw new NullPointerException();
         }
@@ -34,7 +39,7 @@ public class ReservationService {
 
         Reservation newReservation = new Reservation(
                 schedule,
-                reservationRequest.getName()
+                memberName
         );
 
         return reservationDao.save(newReservation);
@@ -49,12 +54,15 @@ public class ReservationService {
         return reservationDao.findAllByThemeIdAndDate(themeId, date);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id, Long memberId) {
         Reservation reservation = reservationDao.findById(id);
+        String memberName = memberDao.findById(memberId).getUsername();
         if (reservation == null) {
             throw new NullPointerException();
         }
-
+        if (!reservation.getName().equals(memberName)){
+            throw new AuthorizationException();
+        }
         reservationDao.deleteById(id);
     }
 }
